fix 5-yolo.py

+++ supervised_learning/object_detection/5-yolo.py
@@ -375,7 +375,7 @@ class Yolo:
             )
 
             # Normalisation [0, 1]
-            normalized = resized.astype(np.float32) / 255.0
+            normalized = resized / 255.0
 
             pimages.append(normalized)
